+++ sentry_sdk/integrations/opentelemetry/integration.py
@@ -9,7 +9,7 @@ from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
 from sentry_sdk.integrations.opentelemetry.span_processor import (
     SentrySpanProcessor,
 )
-from sentry_sdk.sampler import SentrySampler
+from sentry_sdk.opentelemetry.sampler import SentrySampler
 from sentry_sdk.utils import logger
 
 try:
