+++ sentry_sdk/opentelemetry/__init__.py
@@ -1,0 +1,5 @@
+from sentry_sdk.opentelemetry.sampler import SentrySampler
+
+__all__ = [
+    "SentrySampler",
+]
